Add StrayCatServer and StrayCatRunner
StrayCat server runs just like enunu server, start it up and it will print the version and the port it is listening on.

Next use StrayCatRunner as a "resampler" in OpenUtau or UTAU, currently StrayCatRunner is windows only until i can figure out building on for other systems and packaging the libraries into StrayCatRunner so it's seamless to the end user like a normal resampler.

LIBCURL.dll IS NEEDED TO RUN StrayCatRunner!!!!!